🎨 satisfy the mechanical style gate
Split single-line if bodies, move the lease row types to their paired
types file, unnest one ternary, and add the mandatory README sections
to the two new packages.
