+++ libs/backend/agents/execution/protocol/src/prisma-steering-request-unit-of-work.ts
@@ -47,9 +47,15 @@ export class PrismaSteeringRequestUnitOfWork implements SteeringRequestRepositor
 		}
 		catch (error)
 		{
-			if (!___IsRolledBackConflict(error)) throw error;
+			if (!___IsRolledBackConflict(error))
+			{
+				throw error;
+			}
 			const winner = await this._Run(1, function _Read(repository) { return repository.readWinner(command, steeringRequestId); });
-			if (winner !== null) return winner;
+			if (winner !== null)
+			{
+				return winner;
+			}
 			throw error;
 		}
 	}
