+++ libs/backend/agents/execution/runs/main/src/prisma-agent-run-warm-runtime-authority.ts
@@ -428,7 +428,10 @@ export class PrismaAgentRunWarmRuntimeUnitOfWork implements AgentRunWarmRuntimeC
 		}
 		catch (error)
 		{
-			if (!___IsRolledBackConflict(error)) throw error;
+			if (!___IsRolledBackConflict(error))
+			{
+				throw error;
+			}
 			throw new Error("warm runtime reservation conflicted after three attempts", { cause: error });
 		}
 	}
